update configs to show memory limit
